dev-requirements.txt
-r requirements.txt

coverage
codecov
twine
sphinx
sphinx_rtd_theme
sphinx-click
sphinx-autodoc-typehints
pylint
